Revert "Merge branch 'feature/seek-in-logical-segment-bug' into develop"
This reverts commit d8338cd1ee6386f9224e50222e7b1a1fb160f400, reversing
changes made to 8dc99c89580a83852b3ad22ea09444f787092aae.

There is an issue with pause / play events sent when selecting a logical segment. This is not the case on Android and should probably not be. For the moment, these changes are rollbacked and moved to a separate feature branch