Make unit tests involving media playback stable
Since we do not control when events are actually sent by our measurement SDKs (for which network calls are interecepted), we create a separate unit testing session (with a common identifier) for each test, and ensure notification registrations are made in the context of their session. This way we avoid tests conflicting with each other, and associated instabilities.